+++ scripts/profiler.py
@@ -5,7 +5,7 @@ import numpy as np
 from problems.benchmark_problems import get_problem
 from problems.microgrid_function import microgrid_function
 
-from mesh.core import Mesh, MeshParameters
+from amesh.core import AMESH, AMESHParameters
 
 objective_dim = 10
 decision_dim = 30
@@ -60,7 +60,7 @@ func, position_min_value, position_max_value = get_problem('dtlz1', n_obj=object
 #     return microgrid_function(args[0], args[1], args[2], select_bat, load, temperature, solar_data, wind_data)
 
 def run_new():
-    params = MeshParameters(objective_dim, decision_dim,
+    params = AMESHParameters(objective_dim, decision_dim,
                              position_min_value, position_max_value, 
                              population_size,
                              global_guide_method=global_best_attribution_type,
@@ -70,8 +70,8 @@ def run_new():
                              max_personal_guides=personal_guide_array_size,
                              random_state=random_state)
 
-    new_mesh = Mesh(params, func)
-    new_mesh.run()
+    new_amesh = AMESH(params, func)
+    new_amesh.run()
 
 cProfile.run('run_new()', sort='time', filename="profile.prof")
 stats = pstats.Stats('profile.prof')