+++ tests/amesh/validations/test_numpy_validations.py
@@ -3,7 +3,7 @@ from typing import cast
 import numpy as np
 import pytest
 
-from mesh.validations import numpy_validations as npv
+from amesh.validations import numpy_validations as npv
 
 
 def test_assert_np_array_subtype_success():